+++ EditorWindowMain.c
@@ -766,12 +766,12 @@ int ww_getSelectionLines(WINFO* wp, long* pFirstIndex, long* pLastIndex) {
 	*pFirstIndex = -1;
 	*pLastIndex = -1;
 	if (wp->blstart && wp->blend) {
-		*pFirstIndex = ln_indexOf(wp->fp, wp->blstart->lm);
-		if (wp->blstart->lm == wp->blend->lm) {
+		*pFirstIndex = ln_indexOf(wp->fp, wp->blstart->m_linePointer);
+		if (wp->blstart->m_linePointer == wp->blend->m_linePointer) {
 			*pLastIndex = *pFirstIndex;
 		} else {
-			int nDelta = ll_indexOf((LINKED_LIST*)wp->blstart->lm, (LINKED_LIST*)wp->blend->lm);
-			if (wp->blend->lc == 0) {
+			int nDelta = ll_indexOf((LINKED_LIST*)wp->blstart->m_linePointer, (LINKED_LIST*)wp->blend->m_linePointer);
+			if (wp->blend->m_column == 0) {
 				nDelta--;
 			}
 			*pLastIndex = nDelta +*pFirstIndex;